second commit with beginning of a new test and updates to the readme file
