+++ SecretHikeSpotsXCUITests/Screens/HomeScreen.swift
@@ -6,6 +6,6 @@ class HomeScreen: Screen {
     
     static let diamondActionButton: XCUIElement = buttons["plus.diamond.fill, Add new secret hiking spot"]
     static let myHikingSpotsHeader: XCUIElement = staticTexts["My Hiking spots"]
-    static let saveHikingSpotCells: XCUIElement = cells["Near Glasgow, Map pin, Legal"]
+    static let saveHikingSpotCells: XCUIElement = cells["Near Glasvegas, Map pin, Legal"]
     
 }
