--- SecretHikeSpotsXCUITests/Tests/CreateHikingSpotTests.swift
@@ -1,19 +1,0 @@
-
-import Foundation
-import XCTest
-
-class CreateHikingSpotTests: SecretHikeSpotsXCUITestsBase {
-    
-    func testCreateHikingSpot() {
-        givenIHaveAppOpen()
-        whenITapOnAddNewHikingSpotButton()
-        andTheMapPageOpens()
-        andIGiveLocationPermissionIfRequired()
-        andIChooseALocation()
-        andIClickOnSaveThisSpot()
-        andINameMySpot()
-        andIClickSave()
-        ThenIShouldHaveCreatedAHikingSpot()
-    }
-}
-
